+++ packages/@runlightyear/cli/src/shared/execUnsubscribe.ts
@@ -1,9 +1,7 @@
 import getUnsubscribeList from "./getUnsubscribeList";
 import runInContext from "./runInContext";
 import getPreviouslyDeployedCode from "./getPreviouslyDeployedCode";
-import { terminal } from "terminal-kit";
 import uploadUnsubscribeResult from "./uploadUnsubscribeResult";
-import { restoreConsole } from "./restoreConsole";
 import { logDisplayLevel } from "./setLogDisplayLevel";
 import { prepareConsole } from "../logging";
 
@@ -11,7 +9,7 @@ export default async function execUnsubscribe() {
   const compiledCodeStr = await getPreviouslyDeployedCode();
 
   if (!compiledCodeStr) {
-    terminal("No previous code found, skipping unsubscribe");
+    console.debug("No previous code found, skipping unsubscribe step");
     return;
   }
 
@@ -21,7 +19,7 @@ export default async function execUnsubscribe() {
   const unsubscribeList = await getUnsubscribeList();
 
   const doTheUnsubscribe = async (webhookName: string, removed: boolean) => {
-    terminal("Unsubscribing ", webhookName, "\n");
+    console.info("Unsubscribing", webhookName);
 
     const handlerResult = await handler({
       operation: "unsubscribe",
@@ -32,17 +30,13 @@ export default async function execUnsubscribe() {
 
     prepareConsole();
 
-    terminal.gray(
-      "handlerResult",
-      JSON.stringify(handlerResult, null, 2),
-      "\n"
-    );
-
     const { statusCode, body } = handlerResult;
     const responseData = JSON.parse(body);
     const { logs } = responseData;
 
     const status = statusCode >= 300 ? "FAILED" : "SUCCEEDED";
+
+    console.debug("about to upload unsubscribe result");
 
     await uploadUnsubscribeResult({
       webhookName,